+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
+## 0.0.2
+
+- New environment variables and paths are accessible to the running action.
+- Incorrect workflow command parameters fixed.
+- New getters `isDebug` and `getState`.
+
 ## 0.0.1
 
 - Initial version, created by Stagehand